refactor: moved DummyTokenValidator
