+++ wsmaster/che-core-api-user/src/main/java/org/eclipse/che/api/user/server/DummyTokenValidator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2018 Red Hat, Inc.
+ * Copyright (c) 2012-2021 Red Hat, Inc.
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
  * which is available at https://www.eclipse.org/legal/epl-2.0/
@@ -9,12 +9,11 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.api.local;
+package org.eclipse.che.api.user.server;
 
 import javax.inject.Singleton;
 import org.eclipse.che.api.core.ConflictException;
 import org.eclipse.che.api.core.model.user.User;
-import org.eclipse.che.api.user.server.TokenValidator;
 import org.eclipse.che.api.user.server.model.impl.UserImpl;
 
 /**
