+++ wsmaster/pom.xml
@@ -45,7 +45,6 @@
         <module>che-core-api-factory-bitbucket-server</module>
         <module>che-core-api-ssh</module>
         <module>che-core-api-ssh-shared</module>
-        <module>wsmaster-local</module>
         <module>che-core-sql-schema</module>
         <module>che-core-api-system</module>
         <module>che-core-api-system-shared</module>
